Refactor filters code and builder
